Move predicates/matches to their own module
This literally just moves the various predicate and match types. No
code was modified, added, or removed.
