+++ opte/src/engine/arp.rs
@@ -25,18 +25,18 @@ use super::packet::PacketWriter;
 use super::packet::Parsed;
 use super::packet::ReadErr;
 use super::packet::WriteError;
+use super::predicate::ArpHtypeMatch;
+use super::predicate::ArpOpMatch;
+use super::predicate::ArpPtypeMatch;
+use super::predicate::DataPredicate;
+use super::predicate::EtherAddrMatch;
+use super::predicate::EtherTypeMatch;
+use super::predicate::Ipv4AddrMatch;
+use super::predicate::Predicate;
 use super::rule::AllowOrDeny;
-use super::rule::ArpHtypeMatch;
-use super::rule::ArpOpMatch;
-use super::rule::ArpPtypeMatch;
-use super::rule::DataPredicate;
-use super::rule::EtherAddrMatch;
-use super::rule::EtherTypeMatch;
 use super::rule::GenPacketResult;
 use super::rule::HairpinAction;
-use super::rule::Ipv4AddrMatch;
 use super::rule::Payload;
-use super::rule::Predicate;
 use core::convert::TryFrom;
 use core::fmt;
 use core::fmt::Display;
