+++ opte/src/engine/dhcp.rs
@@ -16,15 +16,15 @@ use super::packet::PacketMeta;
 use super::packet::PacketRead;
 use super::packet::PacketReader;
 use super::packet::Parsed;
+use super::predicate::DataPredicate;
+use super::predicate::EtherAddrMatch;
+use super::predicate::IpProtoMatch;
+use super::predicate::Ipv4AddrMatch;
+use super::predicate::PortMatch;
+use super::predicate::Predicate;
 use super::rule::AllowOrDeny;
-use super::rule::DataPredicate;
-use super::rule::EtherAddrMatch;
 use super::rule::GenPacketResult;
 use super::rule::HairpinAction;
-use super::rule::IpProtoMatch;
-use super::rule::Ipv4AddrMatch;
-use super::rule::PortMatch;
-use super::rule::Predicate;
 use super::udp::UdpHdr;
 use super::udp::UdpMeta;
 use core::fmt;
