+++ opte/src/engine/dhcpv6/protocol.rs
@@ -32,15 +32,15 @@ use crate::engine::packet::PacketMeta;
 use crate::engine::packet::PacketRead;
 use crate::engine::packet::PacketReader;
 use crate::engine::packet::Parsed;
+use crate::engine::predicate::DataPredicate;
+use crate::engine::predicate::EtherAddrMatch;
+use crate::engine::predicate::IpProtoMatch;
+use crate::engine::predicate::Ipv6AddrMatch;
+use crate::engine::predicate::PortMatch;
+use crate::engine::predicate::Predicate;
 use crate::engine::rule::AllowOrDeny;
-use crate::engine::rule::DataPredicate;
-use crate::engine::rule::EtherAddrMatch;
 use crate::engine::rule::GenPacketResult;
 use crate::engine::rule::HairpinAction;
-use crate::engine::rule::IpProtoMatch;
-use crate::engine::rule::Ipv6AddrMatch;
-use crate::engine::rule::PortMatch;
-use crate::engine::rule::Predicate;
 use crate::engine::udp::UdpHdr;
 use crate::engine::udp::UdpMeta;
 use core::fmt;
