+++ opte/src/engine/icmp.rs
@@ -15,15 +15,15 @@ use super::packet::PacketMeta;
 use super::packet::PacketRead;
 use super::packet::PacketReader;
 use super::packet::Parsed;
+use super::predicate::DataPredicate;
+use super::predicate::EtherAddrMatch;
+use super::predicate::IpProtoMatch;
+use super::predicate::Ipv4AddrMatch;
+use super::predicate::Predicate;
 use super::rule::AllowOrDeny;
-use super::rule::DataPredicate;
-use super::rule::EtherAddrMatch;
 use super::rule::GenErr;
 use super::rule::GenPacketResult;
 use super::rule::HairpinAction;
-use super::rule::IpProtoMatch;
-use super::rule::Ipv4AddrMatch;
-use super::rule::Predicate;
 use core::fmt;
 use core::fmt::Display;
 pub use opte_api::ip::IcmpEchoReply;
