+++ opte/src/engine/ip4.rs
@@ -32,11 +32,11 @@ use super::headers::RawHeader;
 use super::packet::PacketRead;
 use super::packet::ReadErr;
 use super::packet::WriteError;
-use super::rule::MatchExact;
-use super::rule::MatchExactVal;
-use super::rule::MatchPrefix;
-use super::rule::MatchPrefixVal;
-use super::rule::MatchRangeVal;
+use super::predicate::MatchExact;
+use super::predicate::MatchExactVal;
+use super::predicate::MatchPrefix;
+use super::predicate::MatchPrefixVal;
+use super::predicate::MatchRangeVal;
 pub use opte_api::Ipv4Addr;
 pub use opte_api::Ipv4Cidr;
 pub use opte_api::Ipv4PrefixLen;
