+++ opte/src/engine/nat.rs
@@ -11,12 +11,12 @@ use super::packet::InnerFlowId;
 use super::packet::Packet;
 use super::packet::Parsed;
 use super::port::meta::ActionMeta;
+use super::predicate::DataPredicate;
+use super::predicate::Predicate;
 use super::rule;
 use super::rule::ActionDesc;
 use super::rule::AllowOrDeny;
-use super::rule::DataPredicate;
 use super::rule::HdrTransform;
-use super::rule::Predicate;
 use super::rule::StatefulAction;
 use crate::engine::snat::ConcreteIpAddr;
 use core::fmt;
