+++ opte/src/engine/snat.rs
@@ -19,15 +19,15 @@ use super::packet::Packet;
 use super::packet::PacketMeta;
 use super::packet::Parsed;
 use super::port::meta::ActionMeta;
+use super::predicate::DataPredicate;
+use super::predicate::Predicate;
 use super::rule::ActionDesc;
 use super::rule::AllowOrDeny;
-use super::rule::DataPredicate;
 use super::rule::FiniteResource;
 use super::rule::GenBtError;
 use super::rule::GenDescError;
 use super::rule::GenDescResult;
 use super::rule::HdrTransform;
-use super::rule::Predicate;
 use super::rule::Resource;
 use super::rule::ResourceEntry;
 use super::rule::ResourceError;
