+++ oxide-vpc/src/engine/nat.rs
@@ -33,11 +33,11 @@ use opte::engine::nat::Nat;
 use opte::engine::port::meta::ActionMetaValue;
 use opte::engine::port::PortBuilder;
 use opte::engine::port::Pos;
+use opte::engine::predicate::EtherTypeMatch;
+use opte::engine::predicate::Ipv4AddrMatch;
+use opte::engine::predicate::Ipv6AddrMatch;
+use opte::engine::predicate::Predicate;
 use opte::engine::rule::Action;
-use opte::engine::rule::EtherTypeMatch;
-use opte::engine::rule::Ipv4AddrMatch;
-use opte::engine::rule::Ipv6AddrMatch;
-use opte::engine::rule::Predicate;
 use opte::engine::rule::Rule;
 use opte::engine::snat::NatPool;
 use opte::engine::snat::SNat;
