+++ oxide-vpc/src/engine/overlay.rs
@@ -57,14 +57,14 @@ use opte::engine::port::meta::ActionMeta;
 use opte::engine::port::meta::ActionMetaValue;
 use opte::engine::port::PortBuilder;
 use opte::engine::port::Pos;
+use opte::engine::predicate::DataPredicate;
+use opte::engine::predicate::Predicate;
 use opte::engine::rule::Action;
 use opte::engine::rule::AllowOrDeny;
-use opte::engine::rule::DataPredicate;
 use opte::engine::rule::GenHtError;
 use opte::engine::rule::GenHtResult;
 use opte::engine::rule::HdrTransform;
 use opte::engine::rule::MappingResource;
-use opte::engine::rule::Predicate;
 use opte::engine::rule::Resource;
 use opte::engine::rule::ResourceEntry;
 use opte::engine::rule::Rule;
